refactor: move RowSelectionCheckbox to shared folder
