+++ src/sections/dataset/dataset-files/files-table/FilesTableColumnsDefinition.tsx
@@ -1,12 +1,12 @@
 import { ColumnDef } from '@tanstack/react-table'
 import { FilePreview } from '../../../../files/domain/models/FilePreview'
-import { RowSelectionCheckbox } from './row-selection/RowSelectionCheckbox'
 import { FileInfoCell } from './file-info/file-info-cell/FileInfoCell'
 import { FileInfoHeader } from './file-info/FileInfoHeader'
 import { FileActionsHeader } from './file-actions/FileActionsHeader'
 import { FileActionsCell } from './file-actions/file-actions-cell/FileActionsCell'
 import { FileSelection } from './row-selection/useFileSelection'
 import { FilePaginationInfo } from '../../../../files/domain/models/FilePaginationInfo'
+import { RowSelectionCheckbox } from '@/sections/shared/form/row-selection-checkbox/RowSelectionCheckbox'
 
 export const createColumnsDefinition = (
   paginationInfo: FilePaginationInfo,
